+++ package.json
@@ -18,10 +18,10 @@
     "eslint": "8.38.0",
     "eslint-config-next": "13.3.0",
     "moment": "^2.29.4",
-    "next": "13.3.0",
+    "next": "^13.4.1",
     "next-themes": "^0.2.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sharp": "^0.32.0",
     "typescript": "5.0.4"
   },
