+++ src/components/ui/elongated-image.tsx
@@ -1,3 +1,4 @@
+'use client'
 import Image from 'next/image'
 import { useEffect, useState } from 'react'
 import { ElongatedImageProps as Props } from '@/types/props'
