--- src/pages/_document.tsx
@@ -1,14 +1,0 @@
-import { Html, Head, Main, NextScript } from 'next/document'
-import { useState } from 'react'
-
-export default function Document() {
-  return (
-    <Html lang="en">
-      <Head />
-      <body>
-        <Main />
-        <NextScript />
-      </body>
-    </Html>
-  )
-}
