+++ tailwind.config.js
@@ -3,7 +3,11 @@ import colors from 'tailwindcss/colors'
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   darkMode: 'class',
-  content: ['./src/pages/**/*.tsx', './src/components/**/*.tsx'],
+  content: [
+    './src/pages/**/*.tsx',
+    './src/components/**/*.tsx',
+    './src/app/**/*.tsx',
+  ],
   theme: {
     extend: {
       colors: {
